IOS-1214: Fix landscape-portrait and vice versa issue.
RFC: There are a series of delayed orientation calculations in the dispatch_async and dispatch_delay blocks.  Its best to obtain the current orientation of the device rather than pass down the value through previous outer blocks.  The reason being is that the orientation during the usage may have change or not be the correct/current one at the time its calculations or UI modes have change.

- Encapsulated ‘transformForOrientation()’ method and use calculate transforms based on the immediate orientation value.
- Moved the animation payload. It seems that some of the frame calculations are made prematurely or causing timing issues.
- Stopped passing down the orientation values and started determining that value when regularly needed.